Fix test suite
- Reduced Nz in tests that numerical methods converge to analytical WMT solutions
so that the testing no longer times out; had to make the relative error tolerance
a bit more generous to make it work.

- Update dictionaries of correct answers for example MOM6 data, now with on set of
answers with the `xgcm` method and one set of answers for `xhistogram`.
Note: this is a small dataset so the disagreement between the two looks large;
however, the benchmarks in notebook `4_benchmarking_transformation_methods.ipynb`
suggests that in practical uses the two are virtually indistinguishable.